Single server main function
