--- include/bind_created_socket.h
@@ -1,19 +1,0 @@
-/** @file bind_created_socket.h
- */
-#ifndef bind_created_socket_H
-#define bind_created_socket_H
-
-/**  @brief Function prototypes for binding socket
- *
- *  
- * This function is used to bind socket
- * 
- *
- */
-
-int BindCreatedSocket(int hSocket, int port);
-
-#endif
-
-
-
